REFAC: reorganize project structure to align with hexagonal architecture; move services and repositories to core and secondary layers
