+++ server/src/core/domain/entities/product.entity.ts
