+++ server/src/core/domain/interfaces/secondary/order.repository.ts
@@ -1,4 +1,4 @@
-import { Order } from '../order.entity';
+import { Order } from '../../entities/order.entity';
 
 export interface OrderRepository {
   save(order: Order): Promise<void>;
