+++ server/src/core/domain/interfaces/secondary/product.repository.ts
@@ -1,4 +1,4 @@
-import { Product } from '../product.entity';
+import { Product } from '../../entities/product.entity';
 
 export interface ProductRepository {
   findById(id: string): Promise<Product | null>;
