+++ server/src/infrastructure/secondary/persistence/mongodb/schemas/product.schema.ts
